Refactor code for review
