+++ Assets/Project/Scripts/GameDatas/StageData.cs
@@ -1,4 +1,6 @@
 ﻿using Project.Scripts.Utils.Definitions;
+using Project.Scripts.Utils.PlayerPrefsUtils;
+using System;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -28,5 +30,34 @@ namespace Project.Scripts.GameDatas
         public List<EBulletType> OverviewGimmicks => overviewGimmicks;
 
         public TutorialData Tutorial => tutorial;
+
+        /// <summary>
+        /// ステージのkeyを生成する
+        /// </summary>
+        /// <param name="treeId"></param>
+        /// <param name="stageNumber"></param>
+        /// <returns> StageKey(= treeId_stageNumber) </returns>
+        public static string EncodeStageIdKey(ETreeId treeId, int stageNumber)
+        {
+            return $"{treeId}{PlayerPrefsKeys.KEY_CONNECT_CHAR}{stageNumber}";
+        }
+
+        /// <summary>
+        /// ステージのkeyからステージ情報を返す
+        /// </summary>
+        /// <param name="stageId"></param>
+        /// <returns> (treeId, stageNumber) </returns>
+        public static(ETreeId, int) DecodeStageIdKey(string stageId)
+        {
+            var retValues = stageId.Split(PlayerPrefsKeys.KEY_CONNECT_CHAR);
+            if (retValues.Length != 2) throw new Exception("Wrong key format");
+            try {
+                var treeId = (ETreeId) Enum.ToObject(typeof(ETreeId), retValues[0]);
+                var stageNumber = int.Parse(retValues[1]);
+                return (treeId, stageNumber);
+            } catch (Exception e) {
+                throw e;
+            }
+        }
     }
 }
